+++ CrossTestDocumenter/CrossTestDocumenter.cs
@@ -10,33 +10,38 @@ namespace CrossTestDocumenter
         static void Main(string[] args)
         {
 
-            // Load CrossBreeze CrossTest assembly
-            // Loading from the CrossTestdocumenter debug folder otherwise the getCustomAttribute stepDocAttribute cannot be loaded
-            Assembly configAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\Crosstest_DotNet_SpecFlow\\Lib\\CrossBreeze_CrossTest_SpecFlow_Core\\bin\\Debug\\CrossBreeze.CrossTest.SpecFlow.Core.dll");
-
-            //Assembly databaseConfigAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\CrossTestDocumenter\\bin\\Debug\\CrossBreeze.CrossTest.Database.dll");
-            //Assembly ssisProcessConfigAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\CrossTestDocumenter\\bin\\Debug\\CrossBreeze.CrossTest.Process.Ssis.dll");
-
-            // Get the Config type.
+            // Load CrossBreeze CrossTest Core assembly
+            Assembly configAssembly = Assembly.LoadFrom("C:\\Git\\Repos\\CrossBreeze\\CrossTest\\Crosstest_DotNet_SpecFlow\\Lib\\CrossBreeze_CrossTest_SpecFlow_Core\\bin\\Debug\\CrossBreeze.CrossTest.SpecFlow.Core.dll");
+            // Get the Core Config type.
             Type configType = configAssembly.GetType("CrossBreeze.CrossTest.SpecFlow.Configuration.CrossTestConfig");
-
-            //Type databaseConfigType = configAssembly.GetType("CrossBreeze.CrossTest.Database.Configuration.DatabaseServerConfig");
-            //Type ssisProcessType = configAssembly.GetType("CrossBreeze.CrossTest.SpecFlow.Configuration.CrossTestConfig");
-
             // Create the specflow config documentation.
             new CrossTestConfigDocumenter().CreateConfigDocumentation
             (
                 configType,
-                "C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Generation\\Model\\CrossTestConfig.xml"
+                "C:\\Git\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Generation\\Model\\CrossTestConfig.xml"
             );
 
-            // Create the Specflow Steps documentation.
-            // Replaced by documentation from PDM
-            /*new CrossTestStepsDocumenter().CreateSpecflowDocumentation(
-                configAssembly,
-                "C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Documentation\\generator\\Model\\CrossTestSpec.xml"
+            // Load CrossBreeze CrossTest Ssis assembly
+            Assembly ssisProcessConfigAssembly = Assembly.LoadFrom("C:\\Git\\Repos\\CrossBreeze\\CrossTest\\Crosstest_DotNet_SpecFlow\\Lib\\CrossBreeze_CrossTest_Process_Ssis\\bin\\Debug\\CrossBreeze.CrossTest.Process.Ssis.dll");
+            // Get the Ssis Config type.
+            Type ssisConfigType = ssisProcessConfigAssembly.GetType("CrossBreeze.CrossTest.Process.Ssis.Configuration.crossTestSsisConfig");
+            // Create the specflow config documentation.
+            new CrossTestConfigDocumenter().CreateConfigDocumentation
+            (
+                ssisConfigType,
+                "C:\\Git\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Generation\\Model\\SsisProcessesConfig.xml"
             );
-            */
+
+            // Load CrossBreeze CrossTest Adf assembly
+            Assembly AdfProcessConfigAssembly = Assembly.LoadFrom("C:\\Git\\Repos\\CrossBreeze\\CrossTest\\Crosstest_DotNet_SpecFlow\\Lib\\CrossBreeze_CrossTest_Process_Adf\\bin\\Debug\\CrossBreeze.CrossTest.Process.Adf.dll");
+            // Get the Adf Config type.
+            Type AdfConfigType = AdfProcessConfigAssembly.GetType("CrossBreeze.CrossTest.Process.Adf.Configuration.crossTestAdfConfig");
+            // Create the specflow config documentation.
+            new CrossTestConfigDocumenter().CreateConfigDocumentation
+            (
+                AdfConfigType,
+                "C:\\Git\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Generation\\Model\\AdfProcessesConfig.xml"
+            );
 
         }