+++ Lib/CrossBreeze_CrossTest_SpecFlow_Core/Modules/Data/Database/Context/Context_Helper.cs
@@ -2,6 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using CrossBreeze.CrossTest.Database.Configuration;
 using CrossBreeze.CrossTest.SpecFlow.Helpers;
+using TechTalk.SpecFlow;
 
 namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Context
 {
@@ -9,6 +10,7 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Context
     {
 
         public static void SpecifyTargetDatabaseServer(
+            ScenarioContext scenarioContext,
             string databaseServerName
         )
         {
@@ -22,36 +24,37 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Context
             // Assert the connection string is configured.
             Assert.IsNotNull(defaultConnectionString, string.Format("The connection '{0}' is not configured!", databaseServerName));
             // Set the database connection.
-            DatabaseContext.GetDatabaseContext().SetDefaultDatabaseConnection(serverConfig.Type, defaultConnectionString, serverConfig);
+            DatabaseContext.GetDatabaseContext(scenarioContext).SetDefaultDatabaseConnection(scenarioContext, serverConfig.Type, defaultConnectionString, serverConfig);
 
             // Open a connection to the database server.
-            DatabaseContext.GetDatabaseContext().DatabaseConnection.Open();
+            DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection.Open();
 
             
             
         }
 
         public static void SpecifyTargetDatabase(
+            ScenarioContext scenarioContext,
             string databaseName
         )
         {
             // Before we change the database, the connection must exist and be open.
-            Assert.IsTrue(DatabaseContext.GetDatabaseContext().DatabaseConnection != null, "The default database server connection is not set!");
-            Assert.IsTrue(DatabaseContext.GetDatabaseContext().DatabaseConnection.State == ConnectionState.Open, "The default database connection is not opened!");
+            Assert.IsTrue(DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection != null, "The default database server connection is not set!");
+            Assert.IsTrue(DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection.State == ConnectionState.Open, "The default database connection is not opened!");
 
             // Change the database for the current connection.
-            DatabaseContext.GetDatabaseContext().DatabaseConnection.ChangeDatabase(databaseName);
+            DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection.ChangeDatabase(databaseName);
         }
 
-        public static void SpecifyTestTransaction()
+        public static void SpecifyTestTransaction(ScenarioContext scenarioContext)
         {
             // Before we can begin the transaction, the connection must exist and be open.
-            Assert.IsTrue(DatabaseContext.GetDatabaseContext().DatabaseConnection != null, "The default database server connection is not set!");
-            Assert.IsTrue(DatabaseContext.GetDatabaseContext().DatabaseConnection.State == ConnectionState.Open, "The default database connection is not opened!");
-            Assert.IsNull(DatabaseContext.GetDatabaseContext().DatabaseTransaction, "A transaction on the database connection already exists!");
+            Assert.IsTrue(DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection != null, "The default database server connection is not set!");
+            Assert.IsTrue(DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection.State == ConnectionState.Open, "The default database connection is not opened!");
+            Assert.IsNull(DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseTransaction, "A transaction on the database connection already exists!");
 
             // Begin the transaction with Serializable isolation level.
-            DatabaseContext.GetDatabaseContext().BeginTransaction();
+            DatabaseContext.GetDatabaseContext(scenarioContext).BeginTransaction(scenarioContext);
         }
 
     }