+++ Lib/CrossBreeze_CrossTest_SpecFlow_Core/Modules/Data/Database/DatabaseContextHooks.cs
@@ -11,17 +11,17 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Context
     {
         #region Hooks
         [AfterScenario]
-        public void AfterScenario()
+        public void AfterScenario(ScenarioContext scenarioContext)
         {
             // If the default database server connection is set and open, close the SQL Connection.
-            if (DatabaseContext.GetDatabaseContext().DatabaseConnection != null && DatabaseContext.GetDatabaseContext().DatabaseConnection.State == ConnectionState.Open)
+            if (DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection != null && DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection.State == ConnectionState.Open)
             {
                 // If a transaction exists, do a rollback.
-                if (DatabaseContext.GetDatabaseContext().IsTransactionOpen())
-                    DatabaseContext.GetDatabaseContext().RollbackTransaction();
+                if (DatabaseContext.GetDatabaseContext(scenarioContext).IsTransactionOpen())
+                    DatabaseContext.GetDatabaseContext(scenarioContext).RollbackTransaction();
 
                 // Close the connection.
-                DatabaseContext.GetDatabaseContext().DatabaseConnection.Close();
+                DatabaseContext.GetDatabaseContext(scenarioContext).DatabaseConnection.Close();
             }
         }
         #endregion Hooks