+++ Lib/CrossBreeze_CrossTest_SpecFlow_Core/Modules/Data/Database/Query/Query_Helper.cs
@@ -5,6 +5,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Data;
 using System.IO;
+using TechTalk.SpecFlow;
 using CrossBreeze.CrossDoc.CustomAttributes;
 
 namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Query
@@ -27,20 +28,22 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Query
         /// <param name="queryLanguage">The query language</param>
         /// <param name="queryText">The query text</param>
         /// <returns></returns>
-        private static DataTable ExecuteQueryHelper(QueryLanguage queryLanguage, String queryText)
+        private static DataTable ExecuteQueryHelper(
+            ScenarioContext scenarioContext, QueryLanguage queryLanguage, String queryText)
         {
             switch (queryLanguage)
             {
                 case QueryLanguage.SQL:
-                    return QueryExecutor.ExecuteQueryAndGetResults(DatabaseContext.GetDatabaseContext().CreateDbCommand(), queryText);
+                    return QueryExecutor.ExecuteQueryAndGetResults(DatabaseContext.GetDatabaseContext(scenarioContext).CreateDbCommand(), queryText);
                 case QueryLanguage.MDX:
-                    return QueryExecutor.ExecuteQueryAndGetResults(DatabaseContext.GetDatabaseContext().CreateDbCommand(), queryText);
+                    return QueryExecutor.ExecuteQueryAndGetResults(DatabaseContext.GetDatabaseContext(scenarioContext).CreateDbCommand(), queryText);
                 default:
                     throw new Exception(string.Format("The query language {0} is not supported yet!", Enum.GetName(typeof(QueryLanguage), queryLanguage)));
             }
         }
 
-        private static String GetQueryFileText(String filePath)
+        private static String GetQueryFileText(
+            ScenarioContext scenarioContext, String filePath)
         {
             string directory = Directory.GetParent(Environment.CurrentDirectory).ToString();
             string newDir = Path.GetFullPath(Path.Combine(directory, @"..\"));
@@ -53,6 +56,7 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Query
 
         public static void ExecuteQuery
         (
+            ScenarioContext scenarioContext,
             string queryLanguageName,
             string queryText
         )
@@ -62,21 +66,22 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Data.Database.Query
             Assert.IsNotNull(queryLanguage, string.Format("The given query language is not valid ({0})", queryLanguageName));
 
             // Get the results of the query.
-            DataTable queryResults = ExecuteQueryHelper(queryLanguage, queryText);
+            DataTable queryResults = ExecuteQueryHelper(scenarioContext, queryLanguage, queryText);
             // Assert there are query results.
             Assert.IsNotNull(queryResults, "No query results returned!");
             // Store the results.
-            ResultContext.GetResultContext().SetResultTable(queryResults);
+            ResultContext.GetResultContext(scenarioContext).SetResultTable(queryResults);
         }
 
         
         public static void ExecuteQueryFromFile(
+            ScenarioContext scenarioContext,
             string queryLanguageName,
             string filePath
         )
         {
             // Execute the query.
-            ExecuteQuery(queryLanguageName, GetQueryFileText(filePath));
+            ExecuteQuery(scenarioContext, queryLanguageName, GetQueryFileText(scenarioContext, filePath));
         }
         #endregion