+++ Lib/CrossBreeze_CrossTest_SpecFlow_Core/Modules/Process/Helper/Process_Helper.cs
@@ -3,34 +3,38 @@ using CrossBreeze.CrossTest.SpecFlow.Helpers;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
+using System.Reflection;
 using TechTalk.SpecFlow;
+using CrossBreeze.CrossTest.SpecFlow.Configuration.Process;
 
 namespace CrossBreeze.CrossTest.SpecFlow.Modules.Process
 {
     public static class Process_Helper
     {
 
-        public static void ExecuteProcess(string processName,string processType, string projectName)
+        public static void ExecuteProcess(ScenarioContext scenarioContext, string processName, string processType, string projectName)
         {
             ExecuteProcessHelper(null, processName, null, processType, projectName, null);
         }
 
-        public static void ExecuteProcessWithParameters(string processName, string processType, string projectName, Table parameterTable)
+        public static void ExecuteProcessWithParameters(
+            ScenarioContext scenarioContext, string processName, string processType, string projectName, Table parameterTable)
         {
-            ExecuteProcessHelper(null,processName,null,processType,projectName,parameterTable);
+            ExecuteProcessHelper(null, processName, null, processType, projectName, parameterTable);
         }
 
-        public static void ExecuteTemplatedProcess(string processName, string objectTemplateName, string processType, string projectName)
+        public static void ExecuteTemplatedProcess(ScenarioContext scenarioContext, string processName, string objectTemplateName, string processType, string projectName)
         {
             ExecuteProcessHelper(null, processName, objectTemplateName, processType, projectName, null);
         }
 
-        public static void ExecuteTemplatedProcessWithParameters(string processName, string objectTemplateName, string processType, string projectName, Table parameterTable)
+        public static void ExecuteTemplatedProcessWithParameters(ScenarioContext scenarioContext, string processName, string objectTemplateName, string processType, string projectName, Table parameterTable)
         {
             ExecuteProcessHelper(null, processName, objectTemplateName, processType, projectName, parameterTable);
         }
 
         public static void ExecuteProcessAsRole(
+            ScenarioContext scenarioContext,
             string roleName,
             string processName,
             string processType,
@@ -41,6 +45,7 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Process
         }
 
         public static void ExecuteTemplatedProcessAsRole(
+            ScenarioContext scenarioContext,
             string roleName,
             string processName,
             string objectTemplateName,
@@ -52,6 +57,7 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Process
         }
 
         public static void ExecuteProcessWithParametersAsRole(
+            ScenarioContext scenarioContext,
             string roleName,
             string processName,
             string processType,
@@ -62,6 +68,7 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Process
         }
 
         public static void ExecuteTemplatedProcessAsRoleWithParameters(
+            ScenarioContext scenarioContext,
             string roleName,
             string processName,
             string objectTemplateName,
@@ -71,6 +78,40 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Process
         )
         {
             ExecuteProcessHelper(roleName, processName, objectTemplateName, processType, projectName, parameterTable);
+        }
+
+        public static void ExecuteProcess(string AssemblyName, string className, string methodName, object[] parameterForTheMethod)
+        {
+            MethodInfo mi = null;
+            Type type = null;
+            object responder = null;
+            System.Type[] objectTypes;
+            int count = 0;
+            try
+            {
+                //Load the assembly and get it's information
+                type = System.Reflection.Assembly.Load(AssemblyName).GetType(AssemblyName + "." + className);
+                //Get the Passed parameter types to find the method type
+                objectTypes = new System.Type[parameterForTheMethod.GetUpperBound(0) + 1];
+                foreach (object objectParameter in parameterForTheMethod)
+
+                {
+                    if (objectParameter != null)
+                        objectTypes[count] = objectParameter.GetType();
+                    count++;
+                }
+
+                //Get the reference of the method
+                mi = type.GetMethod(methodName, objectTypes);
+                //ci = type.GetConstructor(Type.EmptyTypes);
+                //responder = ci.Invoke(null);
+                //Invoke the method
+                mi.Invoke(responder, parameterForTheMethod);
+            }
+            catch (Exception ex)
+            {
+                throw ex;
+            }
         }
 
         #region Helpers
@@ -107,7 +148,21 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Process
             }
 
             // Execute the process.
-            ETLHelper.ExecuteProcess((ETLHelper.ProcessType)Enum.Parse(typeof(ETLHelper.ProcessType), processType), projectName, fqProcessName, parameters);
+            var methodParameters = new object[3];
+            methodParameters[0] = projectName;
+            methodParameters[1] = processName;
+            methodParameters[2] = parameters;
+            var processTypeParsed = (Configuration.Process.ProcessType)Enum.Parse(typeof(Configuration.Process.ProcessType), processType);
+            switch (processTypeParsed)
+            {
+                case ProcessType.SSIS:
+                    Process_Helper.ExecuteProcess("CrossBreeze.CrossTest.Process.Ssis", "SsisExecutor", "ExecuteSsisProcess", methodParameters);
+                    break;
+
+                case ProcessType.ADF:
+                    Process_Helper.ExecuteProcess("CrossBreeze.CrossTest.Process.Adf", "AdfExecutor", "ExecuteAdfProcess", methodParameters);
+                    break;
+            }
         }
         #endregion