+++ Lib/CrossBreeze_CrossTest_SpecFlow_Core/Modules/Result/Helper/Result_Helper.cs
@@ -13,20 +13,21 @@ namespace CrossBreeze.CrossTest.SpecFlow.Modules.Result
     public static class Result_Helper
     {
         public static void CompareExpectedAndActualResult(
+            ScenarioContext scenarioContext,
             Table expectedResultsTable
         )
         {
             // The results must be known.
-            Assert.IsNotNull(ResultContext.GetResultContext().ResultTable, "There are no results yet to compare with!");
+            Assert.IsNotNull(ResultContext.GetResultContext(scenarioContext).ResultTable, "There are no results yet to compare with!");
 
             // Get the column types of the result set.
-            Dictionary<String, Type> actualResultTypes = DataHelper.GetColumnTypesFromDataTable(ResultContext.GetResultContext().ResultTable);
+            Dictionary<String, Type> actualResultTypes = DataHelper.GetColumnTypesFromDataTable(ResultContext.GetResultContext(scenarioContext).ResultTable);
 
             // Convert the expected results using the column types from the actual results.
             DataTable expectedDataTable = TableExtensions.ToDataTable(expectedResultsTable, actualResultTypes);
 
             // Derive the delta between the two tables.
-            DataTable differencesTable = DataTableHelper.CompareDataTables(expectedDataTable, ResultContext.GetResultContext().ResultTable);
+            DataTable differencesTable = DataTableHelper.CompareDataTables(expectedDataTable, ResultContext.GetResultContext(scenarioContext).ResultTable);
 
             if (differencesTable != null)
             {